feat: 플레이어 피격 추가, 적 추가, 웨이브 추가 및 수정
- hp, maxhp를 float로 바꿈
- 적 스프라이트와 데이터를 추가, 그에 따른 프리팹 생성
- 새 적들을 웨이브에 배치

+++ Assets/Scripts/Player/PlayerStats.cs
@@ -12,7 +12,7 @@ public class PlayerStats : MonoBehaviour
     public event Action<int> LevelChanged;
     public event Action<float, float> ExpChanged;
     public event Action<int> KillCountChanged;
-    public event Action<int, int> HpChanged;
+    public event Action<float, float> HpChanged;
 
     private int level = 1;
     private float currentExp = 0f;
@@ -39,7 +39,7 @@ public class PlayerStats : MonoBehaviour
     public float MaxExp => maxExp;
     public int KillCount => killCount;
 
-    public int MaxHp => Mathf.Max(MinimumMaxHp, maxHpBase + maxHpBonus);
+    public float MaxHp => Mathf.Max(MinimumMaxHp, maxHpBase + maxHpBonus);
     public int RecoveryAmount => recoveryAmount;
     public int Defense => defense;
     public int ProjectileCountBonus => projectileCountBonus;
@@ -51,21 +51,21 @@ public class PlayerStats : MonoBehaviour
     public float CooldownMultiplier => Mathf.Max(0.1f, 1f - cooldownReductionBonus);
     public float ExpMultiplier => (1f + expBonus);
 
-    public int CurrentHp { get; private set; }
+    public float CurrentHp { get; private set; }
 
     private void Awake()
     {
         CurrentHp = MaxHp;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (damage <= 0)
         {
             throw new System.ArgumentOutOfRangeException(nameof(damage));
         }
 
-        int finalDamage = Mathf.Max(1, damage - Defense);
+        float finalDamage = Mathf.Max(1f, damage - Defense);
 
         SetCurrentHp(CurrentHp - finalDamage);
     }
@@ -110,17 +110,17 @@ public class PlayerStats : MonoBehaviour
     {
         if (value == 0) return;
 
-        int prevMaxHp = MaxHp;
+        float prevMaxHp = MaxHp;
 
         maxHpBonus += value;
-        CurrentHp = Mathf.RoundToInt(CurrentHp * ((float)MaxHp / prevMaxHp));
+        CurrentHp = Mathf.RoundToInt(CurrentHp * (MaxHp / prevMaxHp));
 
         HpChanged?.Invoke(CurrentHp, MaxHp);
     }
 
-    private void SetCurrentHp(int newCurrentHp)
+    private void SetCurrentHp(float newCurrentHp)
     {
-        int prevCurrentHp = CurrentHp;
+        float prevCurrentHp = CurrentHp;
 
         if (newCurrentHp == prevCurrentHp) return;
